refactor(misc): axios to ts
Rename api/src/misc/utils/axios.js to .ts (body verbatim, no annotations
needed) and repoint all importers. No behavior change.
