+++ api/src/datasets/router.js
@@ -13,7 +13,7 @@ import LinkHeader from 'http-link-header'
 import equal from 'deep-equal'
 import slug from 'slugify'
 import * as journals from '../misc/utils/journals.ts'
-import axios from '../misc/utils/axios.js'
+import axios from '../misc/utils/axios.ts'
 import * as esUtils from './es/index.ts'
 import { initDatasetIndex, switchAlias, datasetInfos } from '../datasets/es/manage-indices.js'
 import { computeRealtimeWarnings } from './es/diagnose-warnings.ts'
