+++ api/src/datasets/utils/extensions.ts
@@ -7,7 +7,7 @@ import { Transform, Writable } from 'stream'
 import stringify from 'json-stable-stringify'
 import { flatten } from 'flat'
 import equal from 'deep-equal'
-import axios from '../../misc/utils/axios.js'
+import axios from '../../misc/utils/axios.ts'
 import { fullFilePath, attachmentPath } from './files.ts'
 import { readStreams, writeExtendedStreams } from './data-streams.js'
 import * as restDatasetsUtils from './rest.ts'
