+++ api/src/misc/utils/thumbnails.ts
@@ -6,7 +6,7 @@ import tmp from 'tmp-promise'
 import { Binary } from 'mongodb'
 import config from '#config'
 import mongo from '#mongo'
-import axios from './axios.js'
+import axios from './axios.ts'
 import { setNoCache } from './cache-headers.ts'
 import { tmpDir } from '../../datasets/utils/files.ts'
 import debugLib from 'debug'
