+++ api/src/remote-services/router.js
@@ -4,7 +4,7 @@ import https from 'https'
 import express from 'express'
 import moment from 'moment'
 import slug from 'slugify'
-import axios from '../misc/utils/axios.js'
+import axios from '../misc/utils/axios.ts'
 import pump from '../misc/utils/pipe.ts'
 import CacheableLookup from 'cacheable-lookup'
 import remoteServiceAPIDocs from '../../contract/remote-service-api-docs.js'
