+++ api/src/remote-services/service.ts
@@ -6,7 +6,7 @@ import config from '#config'
 import * as settingsUtils from '../misc/utils/settings.ts'
 import debugLib from 'debug'
 import datasetAPIDocs from '../../contract/dataset-api-docs.ts'
-import axios from '../misc/utils/axios.js'
+import axios from '../misc/utils/axios.ts'
 import { internalError } from '@data-fair/lib-node/observer.js'
 import slug from 'slugify'
 import { type SessionState } from '@data-fair/lib-express'
